requirements.txt
arabic-reshaper==3.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
instagrapi==2.0.0
Pillow==10.0.0
pycryptodomex==3.18.0
pydantic==1.10.9
PySocks==1.7.1
python-bidi==0.4.2
requests==2.31.0
schedule==1.2.0
six==1.16.0
typing_extensions==4.7.1
urllib3==2.0.4
